Adding BLL and DLL Layers for coding the architecture with proper reference to each module to keep a 3-tiered architecture.
