+++ BLL/BLL.iml
@@ -8,5 +8,6 @@
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="module" module-name="DAL" />
+    <orderEntry type="library" name="lib" level="project" />
   </component>
 </module>
